src/main/java/com/github/xtralife/leetcode/problems/p297/serialize_and_deserialize_binary_tree/bfs/Codec.java
package com.github.xtralife.leetcode.problems.p297.serialize_and_deserialize_binary_tree.bfs;

import java.util.LinkedList;
import java.util.Queue;

/**
 * Definition for a binary tree node.
 * public class TreeNode {
 * int val;
 * TreeNode left;
 * TreeNode right;
 * TreeNode(int x) { val = x; }
 * }
 */

public class Codec {
    private static final char delimiter = '|';
    private static final String splitter = "\\|";

    // Encodes a tree to a single string.
    public String serialize(TreeNode root) {
        if (root == null) {
            return "";
        }

        StringBuilder builder = new StringBuilder();
        Queue<TreeNode> queue = new LinkedList<>();
        queue.offer(root);
        while (!queue.isEmpty()) {
            TreeNode node = queue.remove();
            if (node != null) {
                builder.append(node.val);
                queue.offer(node.left);
                queue.offer(node.right);
            }
            builder.append(delimiter);
        }

        return builder.toString();
    }

    // Decodes your encoded data to tree.
    public TreeNode deserialize(String data) {
        if (data.isEmpty()) {
            return null;
        }

        String[] tokens = data.split(splitter);
        TreeNode root = new TreeNode(Integer.parseInt(tokens[0]));
        Queue<TreeNode> queue = new LinkedList<>();
        queue.offer(root);
        for (int i = 1; i < tokens.length; i++) {
            TreeNode node = queue.remove();
            if (!tokens[i].isEmpty()) {
                node.left = new TreeNode(Integer.parseInt(tokens[i]));
                queue.offer(node.left);
            }
            if (++i == tokens.length) {
                break;
            }
            if (!tokens[i].isEmpty()) {
                node.right = new TreeNode(Integer.parseInt(tokens[i]));
                queue.offer(node.right);
            }
        }
        return root;
    }
}

// Your Codec object will be instantiated and called as such:
// Codec ser = new Codec();
// Codec deser = new Codec();
// TreeNode ans = deser.deserialize(ser.serialize(root));